+++ movies/presentation/src/main/java/com/example/movies/presentation/base/BaseFragment.kt
@@ -1,7 +1,5 @@
 package com.example.movies.presentation.base
 
 import androidx.fragment.app.Fragment
-import androidx.navigation.NavDirections
-import androidx.navigation.fragment.findNavController
 
 open class BaseFragment : Fragment()
