+++ movies/presentation/src/main/res/values/strings.xml
@@ -1,1 +1,3 @@
-<resources />
+<resources >
+    <string name="loading">Loading…</string>
+</resources>
